Update algolia indexing trigger (#29)
This workflow cannot be chained from the deploy-main.yaml workflow,
because it will trigger the indexing immediately after the production
branch is updated, which is before the Vercel production deployment is
finished.

The new approach is to run this workflow regularly, and trigger indexing
if there is any recent commits in the production branch.